+++ framework-rest-resources/framework-stream-rest-resources/src/main/java/uk/gov/justice/services/resources/rest/model/StreamResponse.java
@@ -1,0 +1,7 @@
+package uk.gov.justice.services.resources.rest.model;
+
+import java.util.UUID;
+
+public record StreamResponse(UUID streamId, Long position, Long lastKnownPosition, String source, String component,
+                             String updatedAt, boolean upToDate, UUID errorId, Long errorPosition) {
+}