+++ CHANGELOG.md
@@ -7,6 +7,7 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 ### Added
 - New table stream_statistic to include statistics of streams per source, component and state 
 - Added StreamStatisticTimerBean to fill stream_statistic every configured milliseconds
+- Add framework rest endpoint to fetch streams by errorHash
 
 # [17.103.0-M13] - 2025-06-18
 ### Added